chore: move `marcolongo.cloud` `types.ts`
