+++ apps/marcolongo.cloud/src/shared/types.ts
